submit

+++ done/image_content.c
@@ -108,8 +108,8 @@ int lazily_resize(int resolution, struct imgfs_file* imgfs_file, size_t index)
     long end_offset = ftell(imgfs_file->file);
 
 
-    imgfs_file->metadata[index].offset[resolution] = end_offset - buffer_size;
-    imgfs_file->metadata[index].size[resolution] = buffer_size;
+    imgfs_file->metadata[index].offset[resolution] = (uint64_t) ((uint64_t) end_offset - buffer_size);
+    imgfs_file->metadata[index].size[resolution] = (uint32_t) buffer_size;
     imgfs_file->metadata[index].is_valid = 1; // Mark the image as valid
 
     // Writing metadata changes to disk
